requirements.txt
numpy==2.2.6
tensorflow==2.20.0
keras==3.12.0
pandas==2.3.3
fastapi==0.115.5
uvicorn==0.32.1
streamlit==1.51.0
opencv-python==4.10.0.84
pillow==11.0.0
matplotlib==3.9.2




